-> Configured all the testNG tests using groups annotation to execute multiple tests as in a single test-suite.
Description:
‘Groups‘ is one more annotation of TestNG which can be used in the
execution of multiple tests. Let’s say you have hundred tests of class
vehicle and in it ten method of car, ten method of scooter and so on.
You probably like to run all the scooter tests together in a batch. And
you want all to be in a single test suite. With the help of grouping you
can easily overcome this situation.